// components/advanced/src/models/report/ReportView.ts
import { DataResponse } from '../../services/ApiService';

export interface ReportViewResource {
  report_view: ReportView;
}
export interface ReportView {
  items_per_page: number;
  total_items: number;
  columns_headers: string[];
  rows: any[][];
}

export type ReportViewResponse = DataResponse<ReportViewResource>;